requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
python-pptx>=0.6.23
